netrs-ui: modularity
