+++ netrs-ui/src/ui/mod.rs
@@ -1,0 +1,3 @@
+pub mod device_list;
+pub mod sidebar;
+pub mod window;
